[Qt] Allow setting a maximum size for the network cache
Use SQLite freelists to not let the cache grow bigger than
the hard limit.

We use for now a ~5 MB safety margin that might look big for
small caches. In my tests, it effectively limited the cache
to 16 MB, but it won't make much of a difference with large
caches. Bright side is VACUUM is never called and there are
no O(n) queries (like summing the size of all records).

Every time that cache gets close to the hard limit, we simple
drop an arbitrary number of old records (100 for now), acting
like a garbage collector.

Finally, setting a new maximum size bigger than the current cache
size is not supported because that would take a toll on the
normal use case of using always the same maximum size.